extract autocorrection logic to a module
