+++ lib/primer/view_components/linters/label_component_migration_counter.rb
@@ -1,6 +1,7 @@
 # frozen_string_literal: true
 
 require_relative "helpers"
+require_relative "autocorrectable"
 require_relative "argument_mappers/label"
 
 module ERBLint
@@ -8,18 +9,13 @@ module ERBLint
     # Counts the number of times a HTML label is used instead of the component.
     class LabelComponentMigrationCounter < Linter
       include Helpers
+      include Autocorrectable
 
       TAGS = %w[span summary a div].freeze
       CLASSES = %w[Label].freeze
       MESSAGE = "We are migrating labels to use [Primer::LabelComponent](https://primer.style/view-components/components/label), please try to use that instead of raw HTML."
-
-      private
-
-      def map_arguments(tag)
-        ArgumentMappers::Label.new(tag).to_s
-      rescue ArgumentMappers::ConversionError
-        nil
-      end
+      ARGUMENT_MAPPER = ArgumentMappers::Label
+      COMPONENT = "Primer::LabelComponent"
     end
   end
 end
